pyproject.toml
[tool.poetry]
name = "django pro"
version = "0.1.0"
description = ""
authors = ["Ziyodullo ubuntu <ziyodullodev@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.2"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
django-filter = "^23.5"
django-split-settings = "^1.2.0"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pre-commit = "^3.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
